+++ packages/blade-core/src/utils/platform/isReactNative.ts
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/no-namespace */
+import { getPlatformType } from './getPlatformType';
+
+const isReactNative = (): boolean => {
+  return getPlatformType() === 'react-native';
+};
+
+export { isReactNative };