+++ packages/blade-svelte/src/components/Typography/Text/types.ts
@@ -1,0 +1,47 @@
+import type { Snippet } from 'svelte';
+import type { TextAs } from '@razorpay/blade-core/styles';
+import type { StyledPropsBlade } from '@razorpay/blade-core/utils';
+import type { BaseTextProps, BaseTextSizes } from '../BaseText/types';
+
+export type TextVariant = 'body' | 'caption';
+
+type TextCommonProps = {
+  as?: TextAs;
+  truncateAfterLines?: number;
+  children: Snippet | string;
+  weight?: Extract<BaseTextProps['fontWeight'], 'regular' | 'medium' | 'semibold'>;
+  /**
+   * Overrides the color of the Text component.
+   *
+   * **Note** This takes priority over `type` and `contrast` prop to decide color of text
+   */
+  color?: BaseTextProps['color'];
+  textAlign?: BaseTextProps['textAlign'];
+  textTransform?: BaseTextProps['textTransform'];
+  textDecorationLine?: BaseTextProps['textDecorationLine'];
+  wordBreak?: BaseTextProps['wordBreak'];
+  testID?: string;
+} & StyledPropsBlade;
+
+export type TextBodyVariant = TextCommonProps & {
+  variant?: Extract<TextVariant, 'body'>;
+  size?: Extract<BaseTextSizes, 'xsmall' | 'small' | 'medium' | 'large'>;
+};
+
+export type TextCaptionVariant = TextCommonProps & {
+  variant?: Extract<TextVariant, 'caption'>;
+  size?: Extract<BaseTextSizes, 'small' | 'medium'>;
+};
+
+export type TextProps<T> = T extends { variant: infer Variant }
+  ? Variant extends 'caption'
+    ? TextCaptionVariant
+    : Variant extends 'body'
+    ? TextBodyVariant
+    : T
+  : T;
+
+export type GetTextProps<T extends { variant: TextVariant }> = Pick<
+  TextProps<T>,
+  'variant' | 'weight' | 'size' | 'color' | 'testID' | 'textAlign' | 'textDecorationLine'
+>;