+++ packages/blade/src/components/Typography/Code/__tests__/Code.web.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
+import { makeTypographySize } from '~utils/makeTypographySize';
 import { Code } from '../Code';
 import renderWithTheme from '~utils/testing/renderWithTheme.web';
-import { makeTypographySize } from '~utils/makeTypographySize';
 import { makeSpace } from '~utils/makeSpace';
 import { spacing, typography } from '~tokens/global';